Display  company location on a map (#190)
* js for displaying a google map and marker

* feature to show a dynamic Google map for a company

* styles for a company, a dynamic map

* load application setting from shf.yml; set application config value after initialization of other gems and app info

* map partial for displaying a company on a map

* customize the separator for listing categories; html for displaying company info; put company uri into helper (out of a view)

* display map; buttons for actions; add ID and classes for CSS styling; use field_or_none

* geocode company addresses if needed (lazy initialization when fetched from db)

* fix HAML

* style fixes: camelcase, single quotes

* separator = blank space by default

* fix CSS styles per houndci bot

* more CSS fixes per houndcibot

* reorder CSS elements per houndcibot

* sanitize the user entered website

* remove so much nesting in CSS

* add missing semi-colon

* clean up SCSS file

* CSS class and id to surround a view

* add Sanitize gem

* PORO to sanitize an URL string

* only clean up files if Rails is loaded (don't do for tests that didn't load Rails)

* feature: don't let malicious code in the website field be saved when creating or editing a company

* sanitize website parameter; only save if we needed to geocode an address; use flash.now on render

* sanitize the company website before saving so we never put malicious code in the db

* clean up CSS

* update feature to be in line with CSS class cleanup

* Customize our SCSS Lint settings (SelectorDepth is ok at 3 instead of 2)

* use double quotes in SCSS

* put parens around the match method

* turn off MultilinePipe haml-lint rule -- too many false positives

* use leading 0

* needs trailing blankline

* use multi-line comment to make haml-lint happier

+++ spec/spec_helper.rb
@@ -15,6 +15,6 @@ RSpec.configure do |config|
   config.shared_context_metadata_behavior = :apply_to_host_groups
 
   config.after(:suite) do
-    FileUtils.rm_rf(Dir["#{Rails.root}/spec/test_paperclip_files/"])
+    FileUtils.rm_rf(Dir["#{Rails.root}/spec/test_paperclip_files/"]) if Object.const_defined?('Rails')
   end
 end
